move useModalState hook to modal component
