+++ app/client-app/src/components/core/modal.jsx
@@ -7,6 +7,7 @@ import {
   ModalHeader,
   ModalOverlay,
 } from "@chakra-ui/react";
+import { useCallback, useState } from "react";
 import { FaRegCheckCircle, FaTimes } from "react-icons/fa";
 
 export const Modal = ({
@@ -72,3 +73,27 @@ export const Modal = ({
     </ModalContent>
   </ChakraModal>
 );
+
+export const useModalState = (location, navigate, formRef = null) => {
+  const [isModalOpen, setIsModalOpen] = useState(false);
+  const [isLoading, setIsLoading] = useState(false);
+  const [feedback, setFeedback] = useState();
+
+  const handleReset = useCallback(() => {
+    setFeedback();
+    setIsLoading(false);
+    setIsModalOpen(false);
+    formRef?.current?.resetForm();
+    navigate(location.pathname, { replace: true });
+  }, [location.pathname, navigate, formRef]);
+
+  return {
+    isModalOpen,
+    setIsModalOpen,
+    isLoading,
+    setIsLoading,
+    feedback,
+    setFeedback,
+    handleReset,
+  };
+};
